Fix packaging and some code.
