+++ .gitignore
@@ -1,1 +1,4 @@
 bin/*.pdf
+
+*.sw*
+.DS_Store
