+++ Rakefile
@@ -1,0 +1,2 @@
+require_relative 'tasks/assign.rb'
+require_relative 'tasks/train.rb'
