+++ lib/pdf/extract/analysis/columns.rb
