+++ lib/pdf/extract/analysis/margins.rb
