+++ lib/pdf/extract/analysis/sections.rb
