+++ lib/pdf/extract/analysis/titles.rb
