+++ lib/pdf/extract/analysis/zones.rb
