+++ lib/pdf/extract/kmeans.rb
