+++ lib/pdf/extract/model/characters.rb
