+++ lib/pdf/extract/model/chunks.rb
