+++ lib/pdf/extract/model/regions.rb
