+++ lib/pdf/extract/multi_range.rb
