+++ lib/pdf/extract/pdf.rb
