+++ lib/pdf/extract/references/references.rb
