+++ lib/pdf/extract/references/resolve.rb
