+++ lib/pdf/extract/references/resolved_references.rb
