+++ lib/pdf/extract/spatial.rb
