+++ lib/pdf/extract/view/abstract_view.rb
