+++ lib/pdf/extract/view/pdf_view.rb
