+++ lib/pdf/extract/view/xml_view.rb
